+++ libs/clients/consultation-portal/src/lib/consultationPortalClient.module.ts
@@ -1,14 +1,20 @@
 import { Module } from '@nestjs/common'
-import { CasesApi, DocumentsApi, TypesApi } from '../../gen/fetch'
+import { CasesApi, DocumentsApi, TypesApi, UserApi } from '../../gen/fetch'
 
 import {
   DocumentsApiProvider,
   CasesApiProvider,
   TypesApiProvider,
+  UserApiProvider,
 } from './apiConfiguration'
 
 @Module({
-  providers: [DocumentsApiProvider, CasesApiProvider, TypesApiProvider],
-  exports: [CasesApi, DocumentsApi, TypesApi],
+  providers: [
+    DocumentsApiProvider,
+    CasesApiProvider,
+    TypesApiProvider,
+    UserApiProvider,
+  ],
+  exports: [CasesApi, DocumentsApi, TypesApi, UserApi],
 })
 export class ConsultationPortalClientModule {}
